Renamer utils til datoBegrensningUtils.ts for å tydligere utils funksjoner implementert her.
