+++ src/søknader/barnetilsyn/steg/6-barnepass/BarnepassSpørsmål.tsx
@@ -18,7 +18,7 @@ import { harValgtSvar } from '../../../../utils/spørsmålogsvar';
 import { EBarnepass, ETypeBarnepassOrdning, IBarnepassOrdning } from '../../models/barnepass';
 import { EPeriode } from '../../../../models/felles/periode';
 import { DatoBegrensning } from '../../../../components/dato/Datovelger';
-import { erPeriodeGyldigOgInnaforBegrensninger } from '../../../../components/dato/utils';
+import { erPeriodeGyldigOgInnaforBegrensninger } from '../../../../components/dato/datoBegrensningUtils';
 import { Heading, TextField } from '@navikt/ds-react';
 import { SettDokumentasjonsbehovBarn } from '../../../overgangsstønad/models/søknad';
 import { TittelOgSlettKnapp } from '../../../../components/knapper/TittelOgSlettKnapp';
