+++ src/søknader/felles/steg/5-aktivitet/Firma/OmFirmaetDitt.tsx
@@ -8,7 +8,7 @@ import { hentTittelMedNr } from '../../../../../language/utils';
 import { SlettKnapp } from '../../../../../components/knapper/SlettKnapp';
 import styled from 'styled-components';
 import { erStrengGyldigOrganisasjonsnummer } from '../../../../../utils/autentiseringogvalidering/feltvalidering';
-import { erDatoGyldigOgInnaforBegrensninger } from '../../../../../components/dato/utils';
+import { erDatoGyldigOgInnaforBegrensninger } from '../../../../../components/dato/datoBegrensningUtils';
 import { TittelOgSlettKnapp } from '../../../../../components/knapper/TittelOgSlettKnapp';
 import { useLokalIntlContext } from '../../../../../context/LokalIntlContext';
 import { ErrorMessage, Heading, Label, Textarea } from '@navikt/ds-react';
